requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
pydantic-settings>=2.1.0
structlog>=24.1.0
prometheus-client>=0.20.0
pyotp>=2.9.0
passlib[bcrypt]>=1.7.4
pytest>=8.0.0
pytest-asyncio>=0.23.0
